+++ CHANGELOG.md
@@ -3,5 +3,26 @@ All notable changes to the "vscode-sequence-diagrams" extension will be document
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-- Initial release
+## [0.3.0] - 2018-09-15
+### Added
+- Added export to SVG
+- Added export to PNG
+- Added parsing error message is now showing in preview
+
+## [0.2.1] - 2017-09-13
+### Fixed
+- Fixed diagram preview is blank on Unix systems
+
+## [0.2.0] - 2017-09-07
+### Added
+- Added diagram style option to configuration 
+
+![v0.2.0 Perview Screenshot](images/Demo0.2.0.png)
+
+## [0.1.0] - 2017-09-06
+### Added
+- Added preview
+- Added command "Show Sequence Diagram Preview"
+- Added `.seqdiag` extensions support
+
+![v0.1.0 Preview Screenshot](images/Demo0.1.0.png)